% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BTDR-package.R
\docType{package}
\name{BTDR-package}
\alias{BTDR}
\alias{BTDR-package}
\title{Functions to access BUPID Top-Down results}
\description{
This package provides convenient functions to work with results files
generated with BUPID Top-Down. These include conversions to other formats
and common visualizations.
}
\details{
\tabular{ll}{
   Package: \tab BTDR\cr
   Type: \tab Package\cr
   Version: \tab 1.0\cr
   Date: \tab 2014-08-18\cr
   License: \tab What license is it under?\cr
}
}
\author{
Christian Heckendorf

Maintainer: Christian Heckendorf <heckend@bu.edu>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{package}

