% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bupid.R
\docType{class}
\name{bupid-class}
\alias{bupid-class}
\title{Class bupid}
\description{
BUPID Top-Down is a suite of tools designed to analyze top-down
proteomics data.
}
\details{
A \code{bupid} object is an S4 class container that stores the results
of a BUPID Top-Down run and post-processing utilities.
}
\section{Slots}{

\describe{
\item{\code{scan}}{Scan header information}

\item{\code{decon}}{Peak lists}

\item{\code{param}}{Parameters used for each peak list}

\item{\code{prot}}{Protein candidates selected}

\item{\code{search}}{Database search results}

\item{\code{tag}}{Sequence tags identified}

\item{\code{fit}}{Fragment ions identified}

\item{\code{xlink}}{Peak lists with crosslinked ions identified}

\item{\code{xlpep}}{Crosslinked ions identified}
}}
\keyword{Class}

