% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/subset.R
\docType{methods}
\name{bupid-subset}
\alias{bupid-subset}
\alias{subset,bupid-method}
\title{Subset BUPID results.}
\usage{
\S4method{subset}{bupid}(x, subset, select, drop = FALSE)
}
\arguments{
\item{x}{object to be subsetted}

\item{subset}{logical expression indicating elements or rows to keep}

\item{select}{expression, indicating the return type}
}
\value{
Returns the subset
}
\description{
Return a subset of the BUPID results
}
\details{
Filter the results and return items matching the provided conditions.
}
\examples{
server <- "http://bumc-florida.bumc.bu.edu/BUPID_TD/cgi-bin/get_results.cgi"
infile <- "key=WBNqTswT5DPg3aDO&ID=320&date=20150309"
data <- read.bupid(url=paste(server,infile,sep="?"))
subset(data,name=="HBB","fragment")
}

