% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit_error.R
\name{fragment.mass.error}
\alias{fragment.mass.error}
\title{Plot mass errors for fragment assignments.}
\usage{
fragment.mass.error(data)
}
\arguments{
\item{data}{output from read.bupid}
}
\value{
Returns the ggplot object representing the plot.
}
\description{
Generates a plot showing mass error vs. ion mass.
}
\examples{
server <- "http://bumc-florida.bumc.bu.edu/BUPID_TD/cgi-bin/get_results.cgi"
infile <- "key=WBNqTswT5DPg3aDO&ID=320&date=20150309"
data <- read.bupid(url=paste(server,infile,sep="?"))
png(filename="errors.png",width=300,height=100) #Configure a graphics device
fragment.mass.error(data)
dev.off()
}

