% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_tables.R
\name{fragment.matched}
\alias{fragment.matched}
\alias{fragment.matched.clusters}
\alias{fragment.matched.ions}
\alias{fragment.matched.peaks}
\title{Fragment assignment tables}
\usage{
fragment.matched.ions(data)

fragment.matched.peaks(data, format = "list")

fragment.matched.clusters(data)
}
\arguments{
\item{data}{output from read.bupid}

\item{format}{decides the return type: list or matrix}
}
\value{
Returns the assignment information table.

Returns the matrix of assignments.

Returns the list or matrix of assignments.
}
\description{
These functions generate tables using information from the fragment assignment module.

fragment.matched.ions generates a matrix containing detailed information about the assigned fragments.

fragment.matched.peaks generates the assigned fragments with matches grouped by matched peak.

fragment.matched.clusters generates a matrix of assignments along with their
original m/z and charge state values as seen in the raw spectra.
}
\examples{
server <- "http://bumc-florida.bumc.bu.edu/BUPID_TD/cgi-bin/get_results.cgi"
infile <- "key=WBNqTswT5DPg3aDO&ID=320&date=20150309"
data <- read.bupid(url=paste(server,infile,sep="?"))
fragment.matched.ions(data)
fragment.matched.peaks(data)
fragment.matched.clusters(data)
}

