% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spectrum.R
\name{plot.spectrum}
\alias{plot.label.spectrum}
\alias{plot.spectrum}
\alias{plot.theme.simple}
\title{Plot spectra}
\usage{
\method{plot}{spectrum}(data, massrange = c(0, Inf), color = "#000000")

\method{plot}{label.spectrum}(data, massrange = c(0, Inf), unicode = TRUE)

\method{plot}{theme.simple}(base_size = 12)
}
\arguments{
\item{data}{output from read.bupid}

\item{massrange}{a vector describing the lower and upper bounds of the mass range to plot}

\item{color}{the color to use for unlabeled peaks}

\item{unicode}{allow unicode characters in labels}
}
\value{
Returns the ggplot object

Returns the ggplot object

Returns the ggplot object

Returns the ggplot theme
}
\description{
These functions plot the deconvoluted spectra and fragment assignments

plot.spectrum plots the deconvoluted peak list as a mass spectrum

plot.label.spectrum adds fragment assignment labels to plot.spectrum

plot.theme.simple modifies the spectrum plot to use black axis labels and no background
}
\examples{
\dontrun{
server <- "http://bupid.bumc.bu.edu/BUPID_TD/cgi-bin/get_results.cgi"
infile <- "key=WBNqTswT5DPg3aDO&ID=320&date=20150309"
data <- read.bupid(url=paste(server,infile,sep="?"))

# plot the entire spectrum
plot.spectrum(data)

# plot peaks between 500 Da and 600 Da with labeled assignments
plot.label.spectrum(data,c(500,600))

# again with a simple theme
plot.label.spectrum(data,c(500,600)) + plot.theme.simple()
}

}

