% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read.bupid}
\alias{read.bupid}
\title{Read BUPID results.}
\usage{
read.bupid(file = NULL, url = NULL)
}
\arguments{
\item{file}{a file on the local machine to use as input}

\item{url}{a url where the data can be retreived}
}
\value{
Returns the results as a bupid object.
}
\description{
Read BUPID results and create and object from them.
}
\details{
The data can be passed either as the name of a file on the local machine or
a URL on the BUPID server containing the results.
}
\examples{
\dontrun{
server <- "http://bupid.bumc.bu.edu/BUPID_TD/cgi-bin/get_results.cgi"
infile <- "key=WBNqTswT5DPg3aDO&ID=320&date=20150309"
data <- read.bupid(url=paste(server,infile,sep="?"))
}

}

