% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mzidentml.R
\name{write.mzid}
\alias{write.mzid}
\title{Write results as mzIdentML.}
\usage{
write.mzid(data, file, inputfile = NULL)
}
\arguments{
\item{data}{output from read.bupid}

\item{file}{the name of the file you want to save the mzIdentML results as.}

\item{inputfile}{the name of file passed to read.bupid.}
}
\value{
Returns the name of the file generated.
}
\description{
Convert an object created with read.bupid to mzIdentML.
}
\examples{
server <- "http://bumc-florida.bumc.bu.edu/BUPID_TD/cgi-bin/get_results.cgi"
infile <- "key=a1edb86d8f2d4cfcfd065755c5c725f8&ID=136&date=20140801"
data <- read.bupid(url=paste(server,infile,sep="?"))
write.mzid(data,"data.mzid",infile)
}
\seealso{
\code{\link{read.bupid}}
}

