% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlink_table.R
\name{xlinkpep}
\alias{xlinkpep}
\title{Crosslink peptide subset}
\usage{
xlinkpep(object, pepnum)
}
\arguments{
\item{object}{bupid object returned from another function}

\item{pepnum}{peptide number}
}
\value{
Returns a bupid object with results for one peptide
}
\description{
Subset the bupid object by peptide number
}
\examples{
\dontrun{
server <- "http://bupid.bumc.bu.edu/cgi-bin/get_results.cgi"
infile <- "key=WBNqTswT5DPg3aDO&ID=320&date=20150309"
data <- read.bupid(url=paste(server,infile,sep="?"))
sdata <- subset(data,scan.num=234,"protein")
fragment.matched.ions(xlinkpep(sdata,1))
}
}

