% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlink_table.R
\name{xlinks}
\alias{xlinks}
\title{Crosslink results table}
\usage{
xlinks(object, n = 2)
}
\arguments{
\item{object}{bupid object returned from another function}

\item{n}{number of linked peptides to display per scan}
}
\value{
Returns a data.frame with the assignments.
}
\description{
Create a dataframe of xlink results
}
\examples{
\dontrun{
server <- "http://bupid.bumc.bu.edu/cgi-bin/get_results.cgi"
infile <- "key=WBNqTswT5DPg3aDO&ID=320&date=20150309"
data <- read.bupid(url=paste(server,infile,sep="?"))
xlink.table(data,2)
}
}

