% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bupid.R
\docType{class}
\name{bupid-class}
\alias{bupid-class}
\title{Class bupid}
\description{
BUPID Top-Down is a suite of tools designed to analyze top-down
proteomics data.
}
\details{
A \code{bupid} object is an S4 class container that stores the results
of a BUPID Top-Down run and post-processing utilities.
}
\section{Slots}{

\describe{
\item{\code{internal}}{The internal structure used to populate views}

\item{\code{raw}}{Raw results from yaml.load used to populate internal}

\item{\code{view}}{The user-level structure containing a subset of the results}

\item{\code{type}}{The type of view currently active}
}}
\keyword{Class}

