% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/head.R, R/print.R
\docType{methods}
\name{bupid-print}
\alias{bupid-print}
\alias{print,bupid-method}
\title{Display BUPID results.}
\usage{
\S4method{print}{bupid}(x, type = "overview")

\S4method{print}{bupid}(x, type = "overview")
}
\arguments{
\item{type}{The type of output to display}

\item{data}{A bupid object}

\item{data}{A bupid object}

\item{type}{The type of output to display}
}
\value{
Returns the information

Returns the information
}
\description{
BUPID results

BUPID results
}
\details{
Each result from BTDR packages will provide differeny outputs

Each result from BTDR packages will provide differeny outputs
}
\examples{
server <- "http://bumc-florida.bumc.bu.edu/BUPID_TD/cgi-bin/get_results.cgi"
infile <- "key=WBNqTswT5DPg3aDO&ID=320&date=20150309"
data <- read.bupid(url=paste(server,infile,sep="?"))
print(data,"overview")
server <- "http://bumc-florida.bumc.bu.edu/BUPID_TD/cgi-bin/get_results.cgi"
infile <- "key=WBNqTswT5DPg3aDO&ID=320&date=20150309"
data <- read.bupid(url=paste(server,infile,sep="?"))
print(data,"overview")
}

