% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{bupid-summary}
\alias{bupid-summary}
\alias{summary,bupid-method}
\alias{summary.fragment}
\alias{summary.overview}
\alias{summary.protein}
\title{Summarize BUPID results.}
\usage{
\S4method{summary}{bupid}(object, type = "overview")

summary.overview(object, ...)

summary.protein(object, ...)

summary.fragment(object, ...)
}
\arguments{
\item{object}{object returned from another function}

\item{...}{extra parameters}
}
\value{
Returns the summary
}
\description{
Print a summary for BUPID results
}
\details{
Each result from BTDR packages will provide differeny summaries
}
\examples{
\dontrun{
server <- "http://bupid.bumc.bu.edu/cgi-bin/get_results.cgi"
infile <- "key=WBNqTswT5DPg3aDO&ID=320&date=20150309"
data <- read.bupid(url=paste(server,infile,sep="?"))
summary(data)
}

}

