% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_error.R
\name{fragment.mass.error}
\alias{fragment.mass.error}
\title{Plot mass errors for fragment assignments.}
\usage{
fragment.mass.error(data, recalibrate = F)
}
\arguments{
\item{data}{output from read.bupid}

\item{recalibrate}{adjust mass errors so they are centered around zero}
}
\value{
Returns the ggplot object representing the plot.
}
\description{
Generates a plot showing mass error vs. ion mass.
}
\examples{
\dontrun{
server <- "http://bupid.bumc.bu.edu/cgi-bin/get_results.cgi"
infile <- "key=WBNqTswT5DPg3aDO&ID=320&date=20150309"
data <- read.bupid(url=paste(server,infile,sep="?"))
png(filename="errors.png",width=300,height=100) #Configure a graphics device
fragment.mass.error(data)
dev.off()
}

}

