% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit_error.R
\name{fragment.mass.error}
\alias{fragment.mass.error}
\title{Plot mass errors for fragment assignments.}
\usage{
fragment.mass.error(data, fitid = 1L)
}
\arguments{
\item{data}{output from read.bupid}

\item{fitid}{peak fit result index to process}
}
\value{
Returns the ggplot object representing the plot.
}
\description{
Generates a plot showing mass error vs. ion mass.
}
\examples{
infile <- "http://bumc-florida.bumc.bu.edu/BUPID_TD/cgi-bin/get_results.cgi?key=a1edb86d8f2d4cfcfd065755c5c725f8&ID=136&date=20140801"
data <- read.bupid(url=infile)
png(filename="errors.png",width=300,height=100) #Configure a graphics device
fragment.mass.error(data)
dev.off()
}

