% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lcms.R
\name{lcms.proteins}
\alias{lcms.proteins}
\title{Display results from LC-MS/MS data}
\usage{
lcms.proteins(data)
}
\arguments{
\item{data}{output from read.bupid}
}
\value{
Returns a matrix with the results
}
\description{
Display results from LC-MS/MS data processed with BUPID Top-Down
}
\examples{
server <- "http://bumc-florida.bumc.bu.edu/BUPID_TD/cgi-bin/get_results.cgi"
infile <- "key=a1edb86d8f2d4cfcfd065755c5c725f8&ID=136&date=20140801"
data <- read.bupid(url=paste(server,infile,sep="?"))
lcms.proteins(data)
}

