% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spectrum.R
\name{spectrum.plot}
\alias{spectrum.label.plot}
\alias{spectrum.plot}
\alias{spectrum.plot.generic}
\alias{spectrum.theme.simple}
\title{Plot spectra}
\usage{
spectrum.plot.generic(mass, intensity, massrange = c(0, Inf),
  color = "#000000")

spectrum.plot(data, massrange = c(0, Inf), color = "#000000")

spectrum.label.plot(data, massrange = c(0, Inf), unicode = TRUE,
  unassigned = TRUE, mz = FALSE)

spectrum.theme.simple(base_size = 12)
}
\arguments{
\item{massrange}{a vector describing the lower and upper bounds of the mass range to plot}

\item{color}{the color to use for unlabeled peaks}

\item{data}{output from read.bupid}

\item{unicode}{allow unicode characters in labels}

\item{unassigned}{display unassigned peaks behind the labelled spectrum}

\item{mz}{TRUE to plot in m/z, otherwise mass}

\item{base_size}{base text size for axis labels}
}
\value{
Returns the ggplot object

Returns the ggplot object

Returns the ggplot object

Returns the ggplot object

Returns the ggplot theme
}
\description{
These functions plot the deconvoluted spectra and fragment assignments

spectrum.plot.generic plots the deconvoluted peak list as a mass spectrum

spectrum.plot plots the deconvoluted peak list as a mass spectrum

spectrum.label.plot adds fragment assignment labels to spectrum.plot

spectrum.theme.simple modifies the spectrum plot to use black axis labels and no background
}
\examples{
\dontrun{
server <- "http://bupid.bumc.bu.edu/cgi-bin/get_results.cgi"
infile <- "key=WBNqTswT5DPg3aDO&ID=320&date=20150309"
data <- read.bupid(url=paste(server,infile,sep="?"))

# plot the entire spectrum
spectrum.plot(data)

# plot peaks between 500 Da and 600 Da with labeled assignments
spectrum.label.plot(data,c(500,600))

# again with a simple theme
spectrum.label.plot(data,c(500,600)) + spectrum.theme.simple()
}

}

